requirements.txt
# News Aggregator Python Dependencies

# Web Framework
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
starlette==0.26.1
python-multipart==0.0.6
python-jose==3.3.0
PyJWT==2.6.0
passlib==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.7
alembic==1.10.2
psycopg2-binary==2.9.5
asyncpg==0.27.0

# Search
elasticsearch
elasticsearch-dsl

# HTTP and Networking
httpx==0.23.3
aiohttp==3.8.4
requests==2.28.2
beautifulsoup4==4.11.2
feedparser==6.0.10
lxml==4.9.2

# Data Processing
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
nltk==3.8.1
spacy==3.5.1
langdetect==1.0.9
python-dateutil==2.8.2

# Text Processing
markdown==3.4.3
bleach==6.0.0
html2text==2020.1.16
readability-lxml==0.8.1
lxml_html_clean==0.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
schedule==1.1.0
tenacity==8.2.2
cachetools==5.3.0
tqdm==4.65.0

# Testing
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-cov==4.1.0
httpx==0.23.3
pytest-mock==3.10.0

# Development
black==23.1.0
isort==5.12.0
flake8==6.0.0
mypy==1.1.1
pre-commit==3.2.1